requirements.txt
streamlit
pandas
psycopg2-binary
seaborn
matplotlib
altair